Purchase Complete

+++ database/migrations/2026_06_30_111636_create_sales_table.php
@@ -16,12 +16,12 @@ return new class extends Migration
             $table->foreignId('product_id');
             $table->foreignId('customer_id');
             $table->integer('voucher_no')->unique();
-            $table->integer('crate_qty');
-            $table->integer('total_weight');
-            $table->integer('weight_cut')->nullable();
-            $table->integer('netweight');
-            $table->integer('rate');
-            $table->integer('total_amount')->nullable();
+            $table->decimal('crate_qty', 10, 2);
+            $table->decimal('total_weight', 10, 2);
+            $table->decimal('weight_cut', 10, 2)->nullable();
+            $table->decimal('netweight', 10, 2);
+            $table->decimal('rate', 10, 2);
+            $table->decimal('total_amount', 10, 2)->nullable();
             $table->timestamps();
         });
     }
